move fixture to appropriate place
