+++ tests/integration_tests/test_integration_mariadb_example.py
@@ -1,29 +1,9 @@
 import logging
-from typing import Generator
 
 import pytest
 from sqlalchemy import func, select, table
 
-from fidesops.db.session import get_db_session, get_db_engine
-
 logger = logging.getLogger(__name__)
-
-
-@pytest.fixture(scope="session")
-def mariadb_example_db() -> Generator:
-    """Return a connection to the MariaDB example DB"""
-    example_mariadb_uri = (
-        "mariadb+pymysql://mariadb_user:mariadb_pw@mariadb_example/mariadb_example"
-    )
-    engine = get_db_engine(database_uri=example_mariadb_uri)
-    logger.debug(f"Connecting to MariaDB example database at: {engine.url}")
-    SessionLocal = get_db_session(engine=engine)
-    the_session = SessionLocal()
-    # Setup above...
-    yield the_session
-    # Teardown below...
-    the_session.close()
-    engine.dispose()
 
 
 @pytest.mark.integration
